% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogr_public.R
\name{addLayer}
\alias{addLayer}
\title{Add a layer}
\usage{
addLayer(x, lyrname, geomType, srs = NULL, opts = character())
}
\arguments{
\item{x}{a datasource}

\item{lyrname}{the name of the layer to create}

\item{geomType}{a string nameing the stored geometry type}

\item{srs}{a spatial reference system object}

\item{opts}{a vector of strings, each with a KEY=VALUE pair}
}
\description{
Add a new layer to a datasource
}
\examples{
x = newOGRDatasource()
names(x)
addLayer(x, "test", "POLYGON")
names(x)

}
