% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogr_public.R
\name{lapplyFeatures}
\alias{lapplyFeatures}
\alias{lapplyGeometries}
\title{OGR apply-like functions}
\usage{
lapplyFeatures(x, applyFun, ...)

lapplyGeometries(x, applyFun, ...)
}
\arguments{
\item{x}{a data layer}

\item{applyFun}{a function or string}

\item{...}{optional arguments passed to \code{applyFun}}
}
\description{
Apply a function to all features or geometries in a layer
}
\details{
The call signature for is \code{function(feature, ...)} or
\code{function(geometry, ...)}.
}
\examples{
f = system.file("example-data/continents", package = "rgdal2")
x = openOGRLayer(f)
invisible(lapplyFeatures(x, function(a) show(a)))
invisible(lapplyGeometries(x, function(a) show(a)))

}
