% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogr_public.R
\name{as-data-frame}
\alias{as-data-frame}
\title{Dump a layer as a dataframe}
\arguments{
\item{from}{a data layer object}
}
\description{
Converts a layer to a dataframe.
}
\details{
A dataframe is constructed containing all the fields of the data layer.
A geometries are returned in the last column and FIDs in the first column.
The returned dataframe has a special print function to avoid printing all
of the geometries, which could be very large.
}
\examples{
f = system.file("example-data/continents", package = "rgdal2")
x = openOGRLayer(f)
y = as(x, 'data.frame')
show(y)

}
