% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_public.R
\docType{methods}
\name{[[,RGDAL2RasterBand,numeric,numeric-method}
\alias{[[,RGDAL2RasterBand,numeric,numeric-method}
\alias{[[,band}
\alias{[[,RGDAL2RasterBand,missing,missing-method}
\alias{[[,RGDAL2RasterBand,numeric,missing-method}
\alias{[[,RGDAL2RasterBand,missing,numeric-method}
\alias{[[,RGDAL2RasterBand,RGDAL2Geometry,ANY-method}
\title{Subset a raster band}
\usage{
\S4method{[[}{RGDAL2RasterBand,numeric,numeric}(x, i, j)

\S4method{[[}{RGDAL2RasterBand,missing,missing}(x, i, j)

\S4method{[[}{RGDAL2RasterBand,numeric,missing}(x, i, j)

\S4method{[[}{RGDAL2RasterBand,missing,numeric}(x, i, j)

\S4method{[[}{RGDAL2RasterBand,RGDAL2Geometry,ANY}(x, i)
}
\arguments{
\item{x}{a raster band object}

\item{i}{the row indices}

\item{j}{the column indices}
}
\description{
Extract a subset of a raster band and return a new raster band holding the result.
}
\details{
Rather than copying data into R memory, these indexing functions
copy the data to a new dataset. Only the minimum value of \code{i} and
\code{j} and their lengths determine the subset region. If for example
you scramble the values of \code{i} using \code{\link{sample}} (without
replacement) you will get the same result. This is a limitation of the
way that GDAL indexes into raster bands.
}
\examples{
x = newRasterBand(5, 5)
x[[]]
x[] = 1:25
x[[]]
class(x[[]])
x[[1:2]]
x[[c(1, 3), 4:2]]
x[[i = 1:2, j = 3:2]]
e = makeExtent(2, 4, 2, 4)
show(e)
x[[e]]

}
