% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_public.R
\name{copyDataset}
\alias{copyDataset}
\title{Copy a GDAL dataset}
\usage{
copyDataset(x, file = tempfile(), driver = "MEM", opts = character())
}
\arguments{
\item{x}{the source dataset}

\item{file}{the file for the new dataset}

\item{driver}{the driver for the new dataset}

\item{opts}{create options; vector of key=value pairs}
}
\value{
an GDAL dataset object
}
\description{
Copies a source dataset to a new dataset
}
\details{
This function may be convenient for working with a dataset in memory. Note that
not all atrributes are copied at this time. If the driver specified is the same
as the driver of the source dataset and the source dataset was created in
\code{rgdal2}, then the creation options will be passed from the source to the
copy. You can suppress this behavior by setting \code{opts} explicitely, either
to valid options or an invalid option like \code{""}. Invalid options will
generate a warning message, but the dataset will still be copied. The invalid
option strings will propagate to subsequent copies without harm.

The \code{file} parameter is ignored if the \code{driver} parameter
is set to "MEM".
}
\examples{
x = newDataset(10, 10, 3)
y = copyDataset(x)
show(x); show(y)

}
\seealso{
\code{\link{openRasterBand}}, \code{\link{openOGR}}
}
