% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogr_public.R
\docType{methods}
\name{length,RGDAL2Datasource-method}
\alias{length,RGDAL2Datasource-method}
\alias{datasource}
\alias{names,RGDAL2Datasource-method}
\alias{[,RGDAL2Datasource,ANY,ANY-method}
\alias{[[,RGDAL2Datasource,ANY,ANY-method}
\alias{$,RGDAL2Datasource-method}
\title{Methods for data sources}
\usage{
\S4method{length}{RGDAL2Datasource}(x)

\S4method{names}{RGDAL2Datasource}(x)

\S4method{[}{RGDAL2Datasource,ANY,ANY}(x, i)

\S4method{[[}{RGDAL2Datasource,ANY,ANY}(x, i)

\S4method{$}{RGDAL2Datasource}(x, name)
}
\arguments{
\item{x}{the data source}

\item{i}{the layer number}

\item{name}{the name of the layer}
}
\description{
List-like access to layers in a data source
}
\examples{
f = system.file("example-data/continents", package = "rgdal2")
x = openOGR(f)
length(x)
names(x)
class(x[[1]])
class(x[1])
class(x$continent)

}
