% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_public.R
\docType{methods}
\name{dim,RGDAL2Dataset-method}
\alias{dim,RGDAL2Dataset-method}
\alias{dim-dataset}
\alias{dim,RGDAL2RasterBand-method}
\alias{dim-band}
\alias{nband}
\title{Return dimensions of a dataset}
\usage{
\S4method{dim}{RGDAL2Dataset}(x)

\S4method{dim}{RGDAL2RasterBand}(x)

nband(x)
}
\arguments{
\item{x}{a dataset}
}
\description{
Return dimensions of a dataset

Return dimensions of a raster band

Return number of raster bands in a dataset
}
\examples{
f = system.file("example-data/gtopo30_gall.tif", package = "rgdal2")
x = openDataset(f)
dim(x)

f = system.file("example-data/gtopo30_gall.tif", package = "rgdal2")
x = openRasterBand(f)
dim(x)

f = system.file("example-data/bee.jpg", package = "rgdal2")
x = openDataset(f)
nband(x)

}
\seealso{
\code{\link{dim}}

\code{\link{dim}}

\code{\link{dim}}
}
