% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\docType{methods}
\name{draw}
\alias{draw}
\alias{draw,RGDAL2Geometry-method}
\alias{draw,RGDAL2Layer-method}
\alias{draw,RGDAL2RasterBand-method}
\alias{draw,RGDAL2Dataset-method}
\title{Draw a spatial data object}
\usage{
draw(object, ..., region = extent(object), overlay = FALSE,
  recording = TRUE)

\S4method{draw}{RGDAL2Geometry}(object, ..., region = extent(object),
  overlay = FALSE, recording = TRUE)

\S4method{draw}{RGDAL2Layer}(object, ..., region = extent(object),
  overlay = FALSE, recording = TRUE)

\S4method{draw}{RGDAL2RasterBand}(object, ..., dpi = 100,
  region = extent(object), overlay = FALSE, recording = TRUE)

\S4method{draw}{RGDAL2Dataset}(object, ..., dpi = 100,
  region = extent(object), overlay = FALSE, recording = TRUE)
}
\arguments{
\item{object}{the object to draw}

\item{...}{additional arguments passed to \code{\link{grid.draw}}}

\item{region}{a geometry setting the drawing region}

\item{overlay}{if false, clear the graphics device before plotting}

\item{recording}{if false, do not record on the graphics stack}

\item{dpi}{raster image resolution}
}
\value{
the grid graphics object invisibly
}
\description{
Graphically displays the object using the \code{\link{grid}} package
}
\examples{
f = system.file("example-data/bee.jpg", package = "rgdal2")
x = openDataset(f)
draw(x)
  
}
