% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_public.R, R/ogr_public.R
\docType{methods}
\name{extent,RGDAL2Dataset-method}
\alias{extent,RGDAL2Dataset-method}
\alias{extent,RGDAL2RasterBand-method}
\alias{extent,RGDAL2Layer-method}
\alias{extent}
\alias{extent,RGDAL2Geometry-method}
\alias{extent,list-method}
\title{Get the spatial extent of an object}
\usage{
\S4method{extent}{RGDAL2Dataset}(object)

\S4method{extent}{RGDAL2RasterBand}(object)

\S4method{extent}{RGDAL2Layer}(object)

\S4method{extent}{RGDAL2Geometry}(object)

\S4method{extent}{list}(object)
}
\arguments{
\item{object}{a spatial object}
}
\value{
a geometry object
}
\description{
Compute the spatial extent of an object and return it as a geometry
}
\details{
If the object has a spatial filter set, then the extent of the spatial
filter is returned. Otherwise, the extent of the object is returned.
}
\examples{
f = system.file("example-data/gtopo30_gall.tif", package = "rgdal2")
x = openDataset(f)
extent(x)

f = system.file("example-data/gtopo30_gall.tif", package = "rgdal2")
x = openRasterBand(f)
extent(x)

f = system.file("example-data/continents", package = "rgdal2")
x = openOGRLayer(f)
extent(x)

}
