% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_public.R
\name{openDataset}
\alias{openDataset}
\alias{openRasterBand}
\title{Open a GDAL dataset}
\usage{
openDataset(fname, readonly = TRUE, shared = TRUE)

openRasterBand(fname, band = 1L, readonly = TRUE)
}
\arguments{
\item{fname}{the file name}

\item{readonly}{if true, prohibit write operations}

\item{shared}{if true, share the internal GDAL handle}

\item{band}{the band number (1-based)}
}
\value{
\code{openDataset}: a dataset object

\code{openRasterBand}: a raster band
}
\description{
Opens a dataset and returns the dataset or a band from the dataset
}
\details{
A dataset holds raster data organized into multiple layers or bands.
The \code{openRasterBand} function simply opens a dataset and returns
the indicated band. This is a convenience in the common situation of
working with single-band datasets.
}
\examples{
f = system.file("example-data/bee.jpg", package = "rgdal2")
x = openDataset(f)
show(x)
draw(x)

f = system.file("example-data/bee.jpg", package = "rgdal2")
x = openRasterBand(f)
show(x)
draw(x)

}
\seealso{
\code{\link{openRasterBand}}, \code{\link{openOGR}}
}
