% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogr_public.R
\name{\%intersects\%}
\alias{\%intersects\%}
\alias{geometry-binary-ops}
\alias{\%equals\%}
\alias{\%disjoint\%}
\alias{\%touches\%}
\alias{\%crosses\%}
\alias{\%within\%}
\alias{\%contains\%}
\alias{\%overlaps\%}
\alias{\%intersection\%}
\alias{\%union\%}
\alias{\%difference\%}
\alias{\%symdiff\%}
\alias{\%distance\%}
\alias{\%append\%}
\title{Binary geometry functions}
\usage{
lhs \%intersects\% rhs

lhs \%equals\% rhs

lhs \%disjoint\% rhs

lhs \%touches\% rhs

lhs \%crosses\% rhs

lhs \%within\% rhs

lhs \%contains\% rhs

lhs \%overlaps\% rhs

lhs \%intersection\% rhs

lhs \%union\% rhs

lhs \%difference\% rhs

lhs \%symdiff\% rhs

lhs \%distance\% rhs

lhs \%append\% rhs
}
\arguments{
\item{lhs}{a geometry object}

\item{rhs}{a geometry object}
}
\description{
Infix operators for binary geometry functions
}
\details{
These mirror the equivalent functions in OGR. See
\url{http://www.gdal.org/} for more information.
}
\examples{
x = c(0, 0, 1, 1)
y = c(0, 1, 1, 0)
x1 = POLYGON(x = x, y = y)
x2 = POLYGON(x = x + 0.5, y = y + 0.5)

x1 \%intersects\% x2
x1 \%equals\% x2
x1 \%disjoint\% x2
x1 \%touches\% x2
x1 \%crosses\% x2
x1 \%within\% x2
x1 \%contains\% x2
x1 \%overlaps\% x2

x1 \%distance\% x2 

gp1 = grid::gpar(lwd = 6, lty = 2, col = rgb(1, 0, 0, 0.5), fill = NA)
gp2 = grid::gpar(lwd = 6, lty = 2, col = rgb(0, 1, 0, 0.5), fill = NA)
gp3 = grid::gpar(lwd = 6, col = rgb(0, 0, 1, 0.5), fill = rgb(0, 0, 1, 0.25))
roi = extent(x1 \%union\% x2)

x3 = x1 \%intersection\% x2
draw(x1, region = roi, gp = gp1)
draw(x2, gp = gp2, overlay = TRUE)
draw(x3,  gp = gp3, overlay = TRUE)
grid::grid.text("intersection")

x3 = x1 \%union\% x2
draw(x1, region = roi, gp = gp1)
draw(x2, gp = gp2, overlay = TRUE)
draw(x3,  gp = gp3, overlay = TRUE)
grid::grid.text("union")

x3 = x1 \%difference\% x2
draw(x1, region = roi, gp = gp1)
draw(x2, gp = gp2, overlay = TRUE)
draw(x3,  gp = gp3, overlay = TRUE)
grid::grid.text("difference")

x3 = x1 \%symdiff\% x2
draw(x1, region = roi, gp = gp1)
draw(x2, gp = gp2, overlay = TRUE)
draw(x3,  gp = gp3, overlay = TRUE)
grid::grid.text("symmetric difference")

MULTIPOLYGON() \%append\% x1 \%append\% x2

}
