% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogr_public.R
\name{POINT}
\alias{POINT}
\alias{LINESTRING}
\alias{POLYGON}
\alias{MULTIPOINT}
\alias{MULTILINESTRING}
\alias{MULTIPOLYGON}
\alias{GEOMETRYCOLLECTION}
\alias{LINEARRING}
\title{Geometry constructors}
\usage{
POINT(x = 0, y = 0)

LINESTRING(x = c(0, 1), y = c(0, 1))

POLYGON(x = c(0, 0, 1, 1), y = c(0, 1, 1, 0))

MULTIPOINT(x = rnorm(2), y = rnorm(2))

MULTILINESTRING()

MULTIPOLYGON()

GEOMETRYCOLLECTION()

LINEARRING(x = c(0, 0.5, 1), y = c(0, 0.87, 0))
}
\arguments{
\item{x}{x-values}

\item{y}{y-values}
}
\description{
These are shortcuts for creating geometries
}
\examples{
POINT(1, 1)
LINESTRING(c(0, 1), c(1, 0))
POLYGON(c(0, 0, 1, 1, 0), c(0, 1, 1, 0, 0))
MULTIPOINT(c(0, 1), c(0, 1))
mls = MULTILINESTRING()
ls1 = LINESTRING(c(0, 1), c(0, 1))
ls2 = LINESTRING(c(1, 1), c(0, 0))
mls = mls \%append\% ls1
mls \%append\% ls2
mply = MULTIPOLYGON()
mply = mply \%append\% POLYGON()
mply \%append\% POLYGON()
gc = GEOMETRYCOLLECTION()
gc \%append\% mls \%append\% ls1 \%append\% mply
LINEARRING()
POLYGON() \%append\% LINEARRING()

}
