% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogr_public.R
\docType{methods}
\name{length,RGDAL2Geometry-method}
\alias{length,RGDAL2Geometry-method}
\alias{geometry-length}
\alias{[[,RGDAL2Geometry,ANY,ANY-method}
\alias{[[,geometry}
\title{Extract subgeometries}
\usage{
\S4method{length}{RGDAL2Geometry}(x)

\S4method{[[}{RGDAL2Geometry,ANY,ANY}(x, i)
}
\arguments{
\item{x}{a geometry object}

\item{i}{the geometry component index}
}
\description{
Fetch components of a geometry
}
\examples{
f = system.file("example-data/continents", package = "rgdal2")
x = openOGRLayer(f)
length(getNextGeometry(x))

x = newGeometry("MULTIPOINT", list(x = rnorm(3), y = rnorm(3)))
x[[1]]; x[[2]]; x[[3]]

}
