% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogr_public.R
\name{centroid}
\alias{centroid}
\alias{geometry-unary-ops}
\alias{boundary}
\alias{convexHull}
\alias{buffer}
\alias{unionCascaded}
\alias{lineLength}
\alias{area}
\alias{simplify}
\alias{polygonize}
\title{Unary geometry functions}
\usage{
centroid(object)

boundary(object)

convexHull(object)

buffer(object, dist, nseg = 30)

unionCascaded(object)

lineLength(object)

area(object)

simplify(object, tolerance, preserve.topology = TRUE)

polygonize(object)
}
\arguments{
\item{object}{a geometry object}

\item{dist}{buffer distance}

\item{nseg}{number of segements in curve}

\item{tolerance}{larger values increase simplification}

\item{preserve.topology}{if true, attempt to not invalidate the geometry}
}
\description{
These function operate on geometries.
}
\details{
These mirror the equivalent functions in OGR. See
\url{http://www.gdal.org/} for more information.
}
\examples{
x = MULTIPOINT(x = rnorm(100), y = rnorm(100))
y = convexHull(x)
draw(y); draw(x, overlay = TRUE)
centroid(y) 
boundary(y)
unionCascaded(y)
lineLength(LINESTRING(x = c(0, 1), y = c(0, 1)))
area(y)
draw(simplify(y, 2), overlay = TRUE, gp = grid::gpar(fill = NA))
polygonize(x)
draw(buffer(y, 1))
draw(y, overlay = TRUE)

}
