% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{geometryGrob}
\alias{geometryGrob}
\title{Convert a geometry to a grob}
\usage{
geometryGrob(object, ..., units = "native")
}
\arguments{
\item{object}{the geometry}

\item{...}{passed to grid functions}

\item{units}{grid units to use}
}
\description{
Converts the geometry to a grid graphics object
}
\details{
Grid functions operate on grobs. This function generates grobs you can
use in creating grid plots. It is more convenient to use \code{\link{draw}}
as that will do the conversion and build the plot.
}
\examples{
g = graticule()
gg = geometryGrob(g)
show(gg)

}
