% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_public.R
\name{getBand}
\alias{getBand}
\alias{getDataset}
\title{Fetch a raster band object from a dataset}
\usage{
getBand(x, band = 1L)

getDataset(x)
}
\arguments{
\item{x}{a dataset object}

\item{band}{the band number}
}
\value{
a raster band object

a dataset object
}
\description{
Fetch a raster band object from a dataset

Fetch the dataset owning a raster band
}
\details{
Band indices start at 1.

Raster bands are always associated with a dataset and cannot
be deleted. This function fetches the dataset to which the
raster band belongs.
}
\examples{
x = newDataset(100, 100, 3)
show(x); dim(x)
y = getBand(x, 2)
show(y)

x = newRasterBand(100, 100)
show(x)
y = getDataset(x)
show(y)

}
\seealso{
\code{\link{newRasterBand}}, \code{\link{nband}}

\code{\link{getBand}}
}
