% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_public.R
\name{getMetadata}
\alias{getMetadata}
\alias{setMetadata}
\title{Get and set metadata}
\usage{
getMetadata(object, domain = "")

setMetadata(object, metadata, domain = "")
}
\arguments{
\item{object}{a raster object}

\item{domain}{the metadata domain}

\item{metadata}{a vector of key=value strings}
}
\description{
Get and set metadata
}
\examples{
x = newDataset(10, 10)
getMetadata(x)
setMetadata(x, c("key1=value1", "key2=value2"))
getMetadata(x)
getMetadata(x, "private")
setMetadata(x, c("key1=value1", "key2=value2"), "private")
getMetadata(x, "private")

}
