% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_public.R
\name{getNoDataValue}
\alias{getNoDataValue}
\alias{setNoDataValue}
\title{Get or set no data value}
\usage{
getNoDataValue(object)

setNoDataValue(object, no.data.value)
}
\arguments{
\item{object}{a raster band}

\item{no.data.value}{the no data value}
}
\description{
Get or set no data value
}
\details{
If you pass a dataset object, the first band will be used.
\code{setNoDataValue} will return a non-zero value on error.
\code{getNoDataValue} will return NULL if the no data value is
not set.
}
\examples{
f = system.file("example-data/gtopo30_gall.tif", package = "rgdal2")
x = openRasterBand(f)
getNoDataValue(x)
y = copyDataset(x)
getNoDataValue(y)
setNoDataValue(y, -1)
getNoDataValue(y)
f = system.file("example-data/bee.jpg", package = "rgdal2")
x = openDataset(f)
getNoDataValue(x)

}
