% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srs.R
\docType{methods}
\name{getSRS,RGDAL2Dataset-method}
\alias{getSRS,RGDAL2Dataset-method}
\alias{getSRS}
\alias{setSRS}
\alias{getSRS,RGDAL2RasterBand-method}
\alias{getSRS,RGDAL2Geometry-method}
\alias{getSRS,RGDAL2Layer-method}
\alias{setSRS,RGDAL2Geometry,RGDAL2SpatialRef-method}
\alias{setSRS}
\alias{setSRS,RGDAL2Geometry,numeric-method}
\alias{setSRS,RGDAL2Geometry,character-method}
\alias{setSRS,RGDAL2Dataset,RGDAL2SpatialRef-method}
\alias{setSRS,RGDAL2RasterBand,RGDAL2SpatialRef-method}
\alias{setSRS,ANY,NULL-method}
\title{Gets or sets the spatial reference system object}
\usage{
\S4method{getSRS}{RGDAL2Dataset}(object)

\S4method{getSRS}{RGDAL2RasterBand}(object)

\S4method{getSRS}{RGDAL2Geometry}(object)

\S4method{getSRS}{RGDAL2Layer}(object)

\S4method{setSRS}{RGDAL2Geometry,RGDAL2SpatialRef}(object, SRS)

\S4method{setSRS}{RGDAL2Geometry,numeric}(object, SRS)

\S4method{setSRS}{RGDAL2Geometry,character}(object, SRS)

\S4method{setSRS}{RGDAL2Dataset,RGDAL2SpatialRef}(object, SRS)

\S4method{setSRS}{RGDAL2RasterBand,RGDAL2SpatialRef}(object, SRS)

\S4method{setSRS}{ANY,`NULL`}(object, SRS)
}
\arguments{
\item{object}{the object containing or receiving the spatial reference system}

\item{SRS}{the spatial reference system as an object, string or number}
}
\description{
Retrieves or sets the spatial reference system object
}
\note{
Calling \code{setSRS} with \code{SRS = NULL} is a no-op that simply
returns the object without effect.
}
\examples{
f = system.file("example-data/gtopo30_gall.tif", package = "rgdal2")
x = openRasterBand(f)
a = getSRS(x)
show(a)
y = copyDataset(x)
setSRS(y, a)

}
\seealso{
\code{\link{newSRS}}
}
