% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_public.R
\name{getTransform}
\alias{getTransform}
\alias{setTransform}
\alias{copyTransform}
\title{Retrieve the affine transformation}
\usage{
getTransform(object)

setTransform(object, transform)

copyTransform(obj1, obj2)
}
\arguments{
\item{object}{a dataset or raster band}

\item{transform}{a list a returned by \code{\link{getTransform}}}

\item{obj1}{the source dataset or raster band}

\item{obj2}{the target dataset or raster band}
}
\value{
a 2-element list:
\item{transl}{a 2-element vector giving the x and y offsets}
\item{affmat}{a 2x2 scale-rotation matrix}

the dataset invisibly

the target dataset invisibly
}
\description{
Retrieves an offset vector and a rotation matrix that allows
projection from pixel, scan-line (row, column) indices to
geospatial coordinates.

Sets the affine transform coefficients on a dataset. These
allow projection from pixel, scan-line (row, column) indices to
geospatial coordinates.

The affine transform coefficients are copied from one dataset
to another.
}
\examples{
f = system.file("example-data/gtopo30_gall.tif", package = "rgdal2")
x = openDataset(f)
getTransform(x)

f = system.file("example-data/gtopo30_gall.tif", package = "rgdal2")
x = openDataset(f)
y = copyDataset(x)
setTransform(y, getTransform(x))
getTransform(y)

f = system.file("example-data/gtopo30_gall.tif", package = "rgdal2")
x = openDataset(f)
y = copyDataset(x)
copyTransform(x, y)
getTransform(y)

}
