% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_public.R
\name{getBlockSize}
\alias{getBlockSize}
\title{Return the internal blocksize of a dataset}
\usage{
getBlockSize(x)
}
\arguments{
\item{x}{a dataset or raster band}
}
\description{
Return the internal blocksize of a dataset
}
\details{
GDAL datasets can have internal data organized in different ways. Most
datasets are in scanline, pixel (sequential row) arrangement. Other
datasets may be tiled. This allows efficient local access to blocks of
raster data. The GDAL commandline utilities can be used to reblock a
dataset. This function returns the blocksize of the dataset. Block-
aligned access is much more efficient than other data chunking
as GDAL caches whole blocks internally.
}
\note{
GDAL uses and x, y or longitude, latitude convention. The \code{rgdal2}
package will always return values as y, x (row, column) or latitude,
longitude.
}
\examples{
f = system.file("example-data/gtopo30_gall.tif", package = "rgdal2")
x = openDataset(f)
getBlockSize(x)

}
