% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogr_public.R
\name{getFeature}
\alias{getFeature}
\title{Fetch a feature using its identifier}
\usage{
getFeature(x, fid)
}
\arguments{
\item{x}{an OGR layer}

\item{fid}{the feature identifier}
}
\description{
Fetch a feature using its identifier
}
\details{
Each "row" of a data layer is called a feature. Every feature is assigned a unique
identifier or "FID". This function uses the FID as a key into the layer. This function
does not return a geometry. Geometries must be extracted from a feature using
\code{\link{getGeometry}}.
}
\examples{
f = system.file("example-data/continents", package = "rgdal2")
x = openOGRLayer(f)
ids = getIDs(x)
f = getFeature(x, ids[[1]])
show(f)

}
