% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogr_public.R
\name{getFields}
\alias{getFields}
\title{Get the fields of a feature}
\usage{
getFields(x)
}
\arguments{
\item{x}{the feature}
}
\value{
a list of field values
}
\description{
Return fields as R-native objects
}
\examples{
f = system.file("example-data/continents", package = "rgdal2")
x = openOGRLayer(f)
y = getNextFeature(x)
unlist(getFields(y))

}
