% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogr_public.R
\name{getGeometries}
\alias{getGeometries}
\title{Fetch all geometries from a layer}
\usage{
getGeometries(x)
}
\arguments{
\item{x}{a layer object}
}
\description{
Return a vector of geometries
}
\details{
The layer will be \code{\link{rewind}} and all geometries
extracted. After this function is called, the internal feature
cursor will be past the end of the feature list. You must call
\code{\link{rewind}} again before fetching features from the
layer.
}
\examples{
f = system.file("example-data/continents", package = "rgdal2")
x = openOGRLayer(f)
y = getGeometries(x)
draw(extent(x))
invisible(lapply(y,
function(g)
{
 draw(extent(g), overlay = TRUE)
}))

}
