% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogr_public.R
\name{getGeometry}
\alias{getGeometry}
\title{Fetch a geometry from a feature}
\usage{
getGeometry(x)
}
\arguments{
\item{x}{the feature object}
}
\value{
a geometry object
}
\description{
Fetch a geometry from a feature
}
