% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogr_public.R
\name{getGeometryType}
\alias{getGeometryType}
\title{Fetch the geometry type from a feature}
\usage{
getGeometryType(x)
}
\arguments{
\item{x}{the feature object}
}
\value{
a lenght-one character vector
}
\description{
Fetch the geometry type from a feature
}
