% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogr_public.R
\name{getIDs}
\alias{getIDs}
\title{Fetch feature IDs}
\usage{
getIDs(x)
}
\arguments{
\item{x}{a layer object}
}
\description{
Return a list of feature IDs
}
\details{
Every row of a layer has a unique feature id. This function returns them.
Usually this will be a sequence of integers. There is no requirement that
they be contiguous.
}
\examples{
f = system.file("example-data/continents", package = "rgdal2")
x = openOGRLayer(f)
getIDs(x)

}
