% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogr_public.R
\name{getLayerName}
\alias{getLayerName}
\title{Fetch layer name}
\usage{
getLayerName(x)
}
\arguments{
\item{x}{a layer object}
}
\value{
the name of the layer
}
\description{
Fetch layer name
}
\examples{
f = system.file("example-data/continents", package = "rgdal2")
x = openOGRLayer(f)
getLayerName(x)

}
