% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_public.R
\name{getMask}
\alias{getMask}
\title{Fetch the mask associated with a raster band}
\usage{
getMask(x)
}
\arguments{
\item{x}{a raster band or dataset}
}
\value{
a raster mask object
}
\description{
Fetch the mask associated with a raster band
}
\details{
The primary purpose of a mask band is to indicate no-data regions
that should be clipped when drawing and analyzing. A mask band can
also contain alpha values in a RGBA dataset. The meaning of the band
can be ascertained using \code{\link{getMaskFlags}}.

Note that if a dataset does not contain a mask, this function will
still construct and return a mask of all true (non-zero) values.

As a convenience when working with single band datasets, this function
will automatically extract the first band if a dataset is passed. Other
bands are ignored.
}
\examples{
f = system.file("example-data/gtopo30_gall.tif", package = "rgdal2")
x = openRasterBand(f)
show(x)
getMaskFlags(x)
y = getMask(x)
show(y)
draw(y)

}
\seealso{
\code{\link{getMaskFlags}}
}
