% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_public.R
\name{getMaskFlags}
\alias{getMaskFlags}
\title{Fetch flags indicating the mask interpretation}
\usage{
getMaskFlags(x)
}
\arguments{
\item{x}{a raster band, dataset or mask}
}
\value{
\item{no.mask}{true if there is no stored mask}
\item{is.shared}{true if the mask applies to all bands}
\item{is.alpha}{true if the mask contains alpha transparency values}
\item{is.nodata}{true if the maks indicates valid data}
}
\description{
Fetch flags indicating the mask interpretation
}
\note{
As a convenience when working with single band datasets, this function
will automatically extract the first band if a dataset is passed. Other
bands are ignored.
}
\examples{
f = system.file("example-data/gtopo30_gall.tif", package = "rgdal2")
x = openRasterBand(f)
show(x)
getMaskFlags(x)

}
\seealso{
\code{\link{getMask}}
}
