% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogr_public.R
\name{getNextFeature}
\alias{getNextFeature}
\title{Fetch the next feature}
\usage{
getNextFeature(x)
}
\arguments{
\item{x}{an OGR layer}
}
\description{
Returns the next feature in a layer
}
\details{
Each layer holds an internal cursor. This function fetches the next
feature in a sequence and increments the cursor. It can be called
repeatedly to get all the features in a layer.
}
\examples{
f = system.file("example-data/continents", package = "rgdal2")
x = openOGRLayer(f)
g = getNextGeometry(x)
show(g)

}
