% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogr_public.R
\name{getNextGeometry}
\alias{getNextGeometry}
\title{Fetch the next geometry}
\usage{
getNextGeometry(x)
}
\arguments{
\item{x}{an OGR layer}
}
\description{
Returns the next geometry in a layer
}
\details{
Each layer holds an internal cursor. This function fetches the next
geometry in a sequence and increments the cursor. It can be called
repeatedly to get all the geometries in a layer.
}
\examples{
f = system.file("example-data/continents", package = "rgdal2")
x = openOGRLayer(f)
g = getNextGeometry(x)
show(g)

}
