% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogr_public.R
\name{getPoints}
\alias{getPoints}
\title{Get points from geometry}
\usage{
getPoints(x, collapse = FALSE, nested = FALSE)
}
\arguments{
\item{x}{a geometry object}

\item{collapse}{if true, join point lists from all sub-geometries}

\item{nested}{logical; if TRUE, return a more consistently nested list}
}
\description{
Extract a list of x and y coordinates
}
\examples{
getPoints(POLYGON())
getPoints(MULTIPOINT())
getPoints(MULTIPOINT(), collapse = TRUE)

}
