% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srs.R
\name{getPROJ4}
\alias{getPROJ4}
\title{Obtain PROJ.4 representation of an RGDAL2SpatialRef object}
\usage{
getPROJ4(x)
}
\arguments{
\item{x}{object of class RGDAL2SpatialRef}
}
\description{
Obtain PROJ.4 representation of an RGDAL2SpatialRef object
}
