% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{graticule}
\alias{graticule}
\title{Create a graticule}
\usage{
graticule(dlat = 10, dlon = 10, linc = 1)
}
\arguments{
\item{dlat}{the latitude interval in degrees}

\item{dlon}{the longitude interval in degrees}

\item{linc}{the line-segment increment in degrees}
}
\description{
Returns a geometry object containing a graticule
}
\examples{
f = system.file("example-data/continents", package = "rgdal2")
x = openOGRLayer(f)
draw(graticule())
draw(x, gp = grid::gpar(fill = "lightblue"), overlay = TRUE)

}
