% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogr_public.R
\name{featureIter}
\alias{featureIter}
\alias{geometryIter}
\title{Create feature iterator}
\usage{
featureIter(x, reset = TRUE)

geometryIter(x, reset = TRUE)
}
\arguments{
\item{x}{a data layer}

\item{reset}{if true, \code{\link{rewind}} the layer}
}
\description{
Generate a feature iterator for use with \code{\link{foreach}}.
}
\details{
Combined with \code{\link{foreach}} and \code{\%do\%} the
returned object can be used to iterate over all features in a
data layer.
}
\examples{
f = system.file("example-data/continents", package = "rgdal2")
x = openOGRLayer(f)
if (require(foreach)) {
y = foreach(i = featureIter(x),
            .init = data.frame(),
            .combine = rbind) \%do\% getFields(i)
class(y)
show(y)

foreach(i = geometryIter(x, reset = TRUE), .combine = c) \%do\% area(i)
}

}
