% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogr_public.R
\docType{methods}
\name{dim,RGDAL2Layer-method}
\alias{dim,RGDAL2Layer-method}
\alias{vector-layer}
\alias{length,RGDAL2Layer-method}
\alias{names,RGDAL2Layer-method}
\title{Methods for data layers}
\usage{
\S4method{dim}{RGDAL2Layer}(x)

\S4method{length}{RGDAL2Layer}(x)

\S4method{names}{RGDAL2Layer}(x)
}
\arguments{
\item{x}{the data source}
}
\description{
Quasi-dataframe-like access to layers in a data layer (not fully implemented).
}
\examples{
f = system.file("example-data/continents", package = "rgdal2")
x = openOGRLayer(f)
dim(x)
length(x)
names(x)

}
