% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogr_public.R
\name{makeExtent}
\alias{makeExtent}
\title{Generate an extent object}
\usage{
makeExtent(xmin = 0, xmax = 1, ymin = 0, ymax = 1, SRS = NULL)
}
\arguments{
\item{xmin}{left side}

\item{xmax}{right side}

\item{ymin}{bottom}

\item{ymax}{top}

\item{SRS}{a spatial reference system}
}
\value{
a rectanglular polygon geometry
}
\description{
Create an extent based on input boundaries
}
\examples{
show(makeExtent())

}
