% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_public.R
\name{newDataset}
\alias{newDataset}
\title{Create a new GDAL dataset}
\usage{
newDataset(nrow, ncol, nbands = 1L, dataType = "Int32", driver = "MEM",
  file = tempfile(), opts = character(), nosave = FALSE)
}
\arguments{
\item{nrow}{number of rows (scan lines)}

\item{ncol}{number of columns (pixels)}

\item{nbands}{number of bands}

\item{dataType}{the storage data type (see details)}

\item{driver}{the name of the dataset driver}

\item{file}{the name of the file to create}

\item{opts}{dataset creation options (see \url{http://www.gdal.org/formats_list.html})}

\item{nosave}{unlink the file after creation}
}
\value{
an object of class RGDAL2Dataset
}
\description{
Create a new GDAL dataset
}
\details{
For most purposes, only the "GTiff" and "MEM" drivers are needed. "GTiff"
creates a geotiff file-based dataset. The "MEM" driver creates the dataset
in memory and the data will not be saved unless the dataset is copied to a
file-based dataset. Similarly, if \code{nosave} is true, then a dataset will
be created on disk, but the underlying file will be unlinked. This may be useful
for dealing with huge temporary datasets beyond memory capacity. Once a dataset
is closed (this will happen automatically if the dataset object goes out of scope),
all data will be lost unless the dataset is copied to another file-based dataset. 

The data type is given as a string, and can be one of Byte, Int16, Int32,
Float32, Float64. These strings are the same as the GDALDataType enum in the GDAL
distribution, but with the prefix "GDT_" removed. Other data types have limited
support. See \url{http://www.gdal.org} for more information.
}
\note{
Be careful not to unlink an existing file.
}
\examples{
createOpts = c("COMPRESS=LZW", "TILED=YES", "BLOCKXSIZE=16", "BLOCKYSIZE=16")
x = newDataset(64, 64, 5, driver = "GTiff", opts = createOpts)
show(x)
x = newDataset(100, 100, 3, driver = "GTiff", nosave = TRUE)
show(x); dim(x)

}
\seealso{
\code{\link{copyDataset}}

\code{\link{openRasterBand}}, \code{\link{openOGR}}
}
