% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogr_public.R
\name{newGeometry}
\alias{newGeometry}
\title{Create a geometry}
\usage{
newGeometry(geomType, points = NULL, SRS = NULL)
}
\arguments{
\item{geomType}{the name of the geometry type}

\item{points}{an x, y point list}

\item{SRS}{a spatial reference system object}
}
\description{
Creates a geometry object optionally setting the spatial reference system
}
\examples{
newGeometry("MULTIPOINT", list(x = rnorm(3), y = rnorm(3)))

}
\seealso{
\code{\link{POINT}}
}
