% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogr_public.R
\name{newOGRDatasource}
\alias{newOGRDatasource}
\title{Create an OGR datasource}
\usage{
newOGRDatasource(driver = "Memory", fname = tempfile(),
  opts = character())
}
\arguments{
\item{driver}{the file name}

\item{fname}{name of file to use for storage}

\item{opts}{vector of key=value pairs}
}
\value{
an object of class RGDAL2Datasource
}
\description{
Creates a new OGR datasource
}
\details{
The \code{Memory} driver creates the datasource in memeory, in which case
the \code{fname} parameter is ignored.

The \code{rgdal2} package does not yet have much support for creating
and writing layers. This is planned in a future release.
}
\examples{
x = newOGRDatasource()
show(x)

}
