% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_public.R
\name{newRasterBand}
\alias{newRasterBand}
\title{Create a new GDAL raster band}
\usage{
newRasterBand(nrow, ncol, dataType = "Int32", driver = "MEM",
  file = tempfile())
}
\arguments{
\item{nrow}{number of rows (scan lines)}

\item{ncol}{number of columns (pixels)}

\item{dataType}{the storage data type}

\item{driver}{the name of the dataset driver}

\item{file}{the name of the file to create}
}
\value{
a raster band object
}
\description{
This is a convenience wrapper around \code{\link{newDataset}}. It
calls \code{\link{newDataset}} and then returns the first band.
}
\details{
This function does not add a new band to an existing dataset. Very few drivers
support adding bands. Currently that capability is not yet implemented in \code{rgdal2}.
}
\examples{
x = newRasterBand(100, 100)
show(x); dim(x)
y = getDataset(x)
show(y)
  
}
\seealso{
\code{\link{newDataset}}
}
