% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srs.R
\name{newSRS}
\alias{newSRS}
\title{Construct a spatial reference system descriptor}
\usage{
newSRS(defn = "WGS84")
}
\arguments{
\item{defn}{a string defining a spatial reference system}
}
\description{
Builds a spatial reference system descriptor based on a
provided string.
}
\details{
The input definition can be in one of four forms:
1) a PROJ4 string; 2) a Well-Known-Text string; 3) an EPSG code in the
form EPSG:####; or 4) an alias as listed below.

Aliases: \cr
WGS84 = "+proj=longlat +datum=WGS84 +no_defs" \cr
NAD83 = "+proj=longlat +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +no_defs" \cr
USNatAtl = "+proj=laea +lat_0=45 +lon_0=-100 +x_0=0 +y_0=0 +a=6370997 +b=6370997 +units=m +no_defs" \cr
NALCC = "+proj=lcc +lat_1=20 +lat_2=60 +lat_0=40 +lon_0=-96 +x_0=0 +y_0=0 +ellps=GRS80 +datum=NAD83 +units=m +no_defs" \cr
NAAEAC = "+proj=aea +lat_1=20 +lat_2=60 +lat_0=40 +lon_0=-96 +x_0=0 +y_0=0 +ellps=GRS80 +datum=NAD83 +units=m +no_defs" \cr
Robinson = "+proj=robin +lon_0=0 +x_0=0 +y_0=0 +ellps=WGS84 +datum=WGS84 +units=m +no_defs" \cr
Mollweide = "+proj=moll +lon_0=0 +x_0=0 +y_0=0 +ellps=WGS84 +datum=WGS84 +units=m +no_defs" \cr
GRS80 = "+proj=longlat +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +no_defs"
}
\examples{
x = newSRS("EPSG:3126")
show(x)

}
\seealso{
\code{\link{getSRS}}, \code{\link{setSRS}}
}
