% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogr_public.R
\name{openOGR}
\alias{openOGR}
\title{Open an OGR datasource}
\usage{
openOGR(fname, readonly = TRUE)
}
\arguments{
\item{fname}{the file name}

\item{readonly}{if true, disable write operations}
}
\value{
an object of class RGDAL2Datasource
}
\description{
Opens the datasource associated with the specified file.
}
\details{
A data source can be treated as a named list of layers.
}
\examples{
f = system.file("example-data/continents", package = "rgdal2")
x = openOGR(f)
show(x)

}
