% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogr_public.R
\name{openOGRLayer}
\alias{openOGRLayer}
\title{Open an OGR layer}
\usage{
openOGRLayer(fname, layer = 1L, readonly = TRUE)
}
\arguments{
\item{fname}{a file name}

\item{layer}{the layer number}

\item{readonly}{if true, dissable writes}
}
\description{
Opens an OGR data source and returns a layer
}
\examples{
f = system.file("example-data/continents", package = "rgdal2")
x = openOGRLayer(f)
show(x)
draw(x, gp = grid::gpar(fill = "lightblue"))

}
