% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{pickExtent}
\alias{pickExtent}
\title{Pick an extent object}
\usage{
pickExtent(object = NULL, plot.it = TRUE)
}
\arguments{
\item{object}{if not null, assing the output the object's spatial reference system}

\item{plot.it}{if true, overlay the extent on the current plot}
}
\description{
Create an extent geometry by clicking on the plot area
}
