% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogr_public.R
\docType{methods}
\name{properties,RGDAL2Datasource-method}
\alias{properties,RGDAL2Datasource-method}
\alias{properties,RGDAL2Layer-method}
\alias{properties}
\alias{properties,RGDAL2Geometry-method}
\title{Fetch properties of an object}
\usage{
\S4method{properties}{RGDAL2Datasource}(object)

\S4method{properties}{RGDAL2Layer}(object)

\S4method{properties}{RGDAL2Geometry}(object)
}
\arguments{
\item{object}{the object}
}
\value{
a list of properties
}
\description{
Interrogate properties of objects
}
\examples{
f = system.file("example-data/continents", package = "rgdal2")
x = openOGR(f)
properties(x)
x = x[[1]]
properties(x)
x = getNextGeometry(x)
properties(x)

}
