% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srs.R
\docType{methods}
\name{reproject,RGDAL2Dataset,ANY-method}
\alias{reproject,RGDAL2Dataset,ANY-method}
\alias{reproject-dataset}
\alias{reproject,RGDAL2RasterBand,ANY-method}
\alias{reproject-band}
\title{Simple raster warping}
\usage{
\S4method{reproject}{RGDAL2Dataset,ANY}(object, SRS, file = tempfile(),
  driver = "MEM", thresh = 0.125)

\S4method{reproject}{RGDAL2RasterBand,ANY}(object, SRS, file = tempfile(),
  driver = "MEM", thresh = 0.125)
}
\arguments{
\item{object}{the raster object (dataset or band)}

\item{SRS}{the new spatial reference system}

\item{file}{the output filed (ignored if \code{driver = "MEM"})}

\item{driver}{the output driver}

\item{thresh}{resampling threshold}
}
\description{
Reproject a raster object
}
\details{
This function calls a modified version of the GDAL warpsimple utility.

If \code{object} is a raster band, the entire dataset will be warped
and the first band of the dataset returned. See \code{\link{getDataset}}.
}
\examples{
f = system.file("example-data/gtopo30_gall.tif", package = "rgdal2")
x = openDataset(f)
show(grid::grid.layout(1, 2))
draw(x)
y = reproject(x, newSRS("WGS84"))
draw(y)

}
