% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srs.R
\docType{methods}
\name{reproject,RGDAL2Geometry,RGDAL2SpatialRef-method}
\alias{reproject,RGDAL2Geometry,RGDAL2SpatialRef-method}
\alias{reproject-geometry}
\alias{reproject,RGDAL2Geometry,numeric-method}
\alias{reproject,RGDAL2Geometry,character-method}
\alias{reproject,ANY,NULL-method}
\title{Transform vector geometries}
\usage{
\S4method{reproject}{RGDAL2Geometry,RGDAL2SpatialRef}(object, SRS)

\S4method{reproject}{RGDAL2Geometry,numeric}(object, SRS)

\S4method{reproject}{RGDAL2Geometry,character}(object, SRS)

\S4method{reproject}{ANY,`NULL`}(object, SRS)
}
\arguments{
\item{object}{the object to reproject}

\item{SRS}{a spatial reference system descriptor}
}
\description{
Project data to new reference system
}
\details{
The SRS can be given as an SRS object, a string or number. Numbers will be coverted to an EPSG code.
Strings can be SRS aliases (see \code{\link{newSRS}}) or anything GDAL understands including PROJ4
strings.
}
\examples{
g1 = newGeometry("POINT", list(x = 1, y = 1), newSRS("WGS84"))
show(g1)
g2 = reproject(g1, newSRS("Moll"))
show(g2)

f = system.file("example-data/continents", package = "rgdal2")
x = openOGRLayer(f)

g = getGeometries(x)
draw(reproject(graticule(), getSRS(x)))
invisible(lapply(g, draw, overlay = TRUE, gp = grid::gpar(fill = "lightblue")))
grid::grid.text("WGS84 projection (lon-lat)", y = 1.1)
   
g = lapply(g, reproject, "Robinson")
draw(reproject(graticule(), "Robinson"))
invisible(lapply(g, draw, overlay = TRUE, gp = grid::gpar(fill = "lightblue")))
grid::grid.text("Robinson projection", y = 1.1)

}
