% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defs.R
\name{reproject}
\alias{reproject}
\title{Reproject an object}
\usage{
reproject(object, SRS, ...)
}
\arguments{
\item{object}{the object to reproject}

\item{SRS}{the new spatial reference system}

\item{...}{additional arguments}
}
\description{
Convert object to a new spatial reference system
}
