% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogr_public.R
\name{rewind}
\alias{rewind}
\title{Rewind OGR layer}
\usage{
rewind(x)
}
\arguments{
\item{x}{an OGR layer}
}
\description{
Reset the geometry cursor to the beginning
}
\details{
The function \code{\link{getNextGeometry}} will increment forward until
no geometries remain. To start with the first geometry, this function must
be called if the cursor is not already at the begining.
}
