% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{setViewport}
\alias{setViewport}
\title{Set the grid viewport}
\usage{
setViewport(object, ..., recording = TRUE)
}
\arguments{
\item{object}{any object for which \code{\link{extent}} is defined}

\item{...}{additional arguments passed to \code{\link{viewport}}}

\item{recording}{if true, grid will record a copy of the grob}
}
\description{
Uses the \code{\link{extent}} of a spatial object to
set the viewport.
}
