% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogr_public.R
\name{setSpatialFilter}
\alias{setSpatialFilter}
\alias{getSpatialFilter}
\alias{setSelectWhere}
\title{Set a spatial filter on a data layer}
\usage{
setSpatialFilter(layer, geom)

getSpatialFilter(layer)

setSelectWhere(layer, where)
}
\arguments{
\item{layer}{a data layer}

\item{geom}{a geometry}

\item{where}{an SQL where statement}
}
\description{
Constrains the data layer by the geometry
}
\details{
This should in principle cause \code{\link{getGeometries}} to only
return those that intersect the filter geometry. Its not clear to me
that this actually works.
}
\examples{
f = system.file("example-data/continents", package = "rgdal2")
x = openOGRLayer(f)
setSelectWhere(x, "\\'CONTINENT\\' = \\'Africa\\'")
# g = getNextGeometry(x);  # EP: breaks
rewind(x) 
# as(x, 'data.frame') #  EP: breaks
setSelectWhere(x, "")
as(x, 'data.frame')
\dontrun{
setSpatialFilter(x, g)
as(x, 'data.frame')
draw(getSpatialFilter(x))
}

}
