% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogr_public.R
\name{getSQLLayer}
\alias{getSQLLayer}
\alias{execSQL}
\title{Query data source and return a layer}
\usage{
getSQLLayer(x, sql)

execSQL(x, sql)
}
\arguments{
\item{x}{a data source}

\item{sql}{an SQL query string}
}
\description{
Execute SQL query and return results as a layer
}
\details{
\code{getSQLLayer} executes the SQL query and returns the results as a layer,
whereas \code{execSQL}, as a convenience, simply executes the SQL and discards
the result. Note that a layer could be simply a table of values without any
associated geometries.
}
\examples{
f = system.file("example-data/continents", package = "rgdal2")
x = openOGR(f)
y = getSQLLayer(x, "select * from continent where \\'CONTINENT\\' = \\'Australia\\'")
# show(y) # EP: breaks
# draw(y, region = getNextGeometry(y)) # EP: breaks;

execSQL(x, "select * from continent")  #only for side-effect

}
