% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogr_public.R
\docType{methods}
\name{getSRS,RGDAL2Feature-method}
\alias{getSRS,RGDAL2Feature-method}
\title{Get the spatial reference system from a feature}
\usage{
\S4method{getSRS}{RGDAL2Feature}(object)
}
\arguments{
\item{object}{the object holding the SRS}
}
\description{
Get the spatial reference system from a feature
}
