% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_public.R
\docType{methods}
\name{[,RGDAL2RasterBand,numeric,numeric-method}
\alias{[,RGDAL2RasterBand,numeric,numeric-method}
\alias{[,band}
\alias{[,RGDAL2RasterBand,missing,missing-method}
\alias{[,RGDAL2RasterBand,numeric,missing-method}
\alias{[,RGDAL2RasterBand,missing,numeric-method}
\alias{[,RGDAL2RasterBand,RGDAL2Geometry,missing-method}
\title{Extract data from a raster band}
\usage{
\S4method{[}{RGDAL2RasterBand,numeric,numeric}(x, i, j, ..., drop = TRUE)

\S4method{[}{RGDAL2RasterBand,missing,missing}(x, i, j, ..., drop = TRUE)

\S4method{[}{RGDAL2RasterBand,numeric,missing}(x, i, j, ..., drop = TRUE)

\S4method{[}{RGDAL2RasterBand,missing,numeric}(x, i, j, ..., drop = TRUE)

\S4method{[}{RGDAL2RasterBand,RGDAL2Geometry,missing}(x, i, j, ...,
  drop = TRUE)
}
\arguments{
\item{x}{a raster band}

\item{i}{the row indices}

\item{j}{the column indicies}

\item{...}{additional arguments (see details)}

\item{drop}{if true (default), drop singleton dimensions}
}
\description{
Use R-style syntax to read data from a raster band
}
\details{
A raster band emulates an R array. Indexing should operate similarly
to native R objects. GDAL internal access functions only allow fetching
of square blocks of data, so \code{x[c(1, 100), c(1, 100)]} will have the
same overhead as \code{x[1:100, 1:100]}. The first however will return a
2x2 matrix whereas the second will return a 100x100 matrix. If either \code{i}
or \code{j} are missing, they assume the values \code{1:nrow(x)} and
\code{1:ncol(x)} respectively.

The subsampling indices \code{ii} and \code{jj} can be used to subsample
the extracted data. The output matrix will have \code{length(ii)} rows and
\code{length(jj)} columns. If the number of elements of \code{ii} (\code{jj})
is smaller than the number of elements in \code{i} (\code{j}) then the
image will be subsampled by skipping pixel values. If the opposite is true,
pixel values will be duplicated to fill out the returned matrix. Note that
whenever the subsampling indices are applied, the minimum value is subtracted
so that \code{ii = 1:3} is equivalent to \code{ii = 101:103}.

If \code{use.mask} is set to \code{FALSE}, then all data will be read. Otherwise
data values where the mask is false will be returned as \code{NA}.

As a special case, \code{i} may be given as a geometry. The extent of the
geometry will be extracted and used to subset the raster data. The geometry
will be reprojected to match the spatial reference system  of the dataset.
This can be combined with the \code{ii} and \code{jj} parameters.

Data are always returned as R type \code{numeric}. The functions
\code{\link{readBlock}} and \code{\link{writeBlock}} are faster
and will return raw bytes, integer or numeric values depending on the type
of data stored in the band.
}
\examples{
x = newRasterBand(5, 5)
x[]
x[] = 1:25
x[]
x[1:2]
x[c(1, 3), 4:2]
x[ii = 1:3, jj = c(4, 2)]
x[i = 1:2, j = 3:2, ii = 1:5, jj = 1:7]
e = makeExtent(2, 4, 2, 4)
show(e)
x[e]

}
\seealso{
\code{\link{readBlock}}
}
