% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_public.R
\docType{methods}
\name{[<-,RGDAL2RasterBand,missing,missing-method}
\alias{[<-,RGDAL2RasterBand,missing,missing-method}
\alias{[<-,band}
\alias{[<-,RGDAL2RasterBand,numeric,missing-method}
\alias{[<-,RGDAL2RasterBand,missing,numeric-method}
\alias{[<-,RGDAL2RasterBand,numeric,numeric-method}
\title{Write data to a raster band}
\usage{
\S4method{[}{RGDAL2RasterBand,missing,missing}(x, i, j, ...) <- value

\S4method{[}{RGDAL2RasterBand,numeric,missing}(x, i, j, ...) <- value

\S4method{[}{RGDAL2RasterBand,missing,numeric}(x, i, j, ...) <- value

\S4method{[}{RGDAL2RasterBand,numeric,numeric}(x, i, j, ...) <- value
}
\arguments{
\item{x}{a raster band}

\item{i}{the row indices}

\item{j}{the column indicies}

\item{...}{optional arguments (see details)}

\item{value}{the data to write}
}
\description{
Use R-style semantics to write data to a raster band
}
\details{
In GDAL, the rows and columns of the data written to a raster band can be
different that the rows and columns touched by the write. The input will
be exanded or subsampled to fit the area written. Hence the range of \code{i}
and \code{j} need not be exactly the size of the data written to the band.

If \code{native.indexing} is true, then a region of raster band data will be
read from the band first. The assignment will then take place on this R matrix
and the matrix written back to the band. This will be slower, but allows one
to write to non-contiguous indexes as in R. There is a performance penalty
as the data block needs to be read and written back to the band.

If the subsetting indices \code{ii} or \code{jj} are set and \code{native.indexing}
is false, then their range of values will be interpreted as the matrix dimensions of
\code{value}. When \code{native.indexing} is true, these arguments are ignored.

This method of writting data is much slower than \code{\link{writeBlock}}
but will automatically cast the data to the correct type unlike \code{\link{writeBlock}}
which requires the input to be of the correct \code{\link{storage.mode}}.
}
\examples{
x = newRasterBand(5, 5); x[]
x[] = 1:25; x[]
x[2:3, 2:3] = matrix(101:125, 5, 5); x[]
x[2:4, 2:4] = matrix(101:125, 5, 5); x[]
x[] = 1; x[]
x[c(2, 4), c(2, 4)] = 2:10; x[]; x[] = 1
x[c(2, 4), c(2, 4), native.indexing = TRUE] = 2:5; x[]
x[ii = 1:3, jj = 1:3] = 1:9; x[]

}
\seealso{
\code{\link{writeBlock}}
}
